+++ apps/backend-functions/src/internals/invitations/events.spec.ts
@@ -3,6 +3,7 @@ import { getCollectionRef, loadAdminServices } from "@blockframes/firebase-utils
 import { getTestingProjectId, initFunctionsTestMock, populate } from "@blockframes/testing/firebase/functions";
 import { clearFirestoreData } from "@firebase/testing";
 import { createNotificationsForEventsToStart } from './events';
+import * as data from "@blockframes/testing/fixtures/events";
 
 describe('Events tests', () => {
 
@@ -17,65 +18,11 @@ describe('Events tests', () => {
   });
 
   it('should create one notification to users with accepted invitation if events are about to start', async () => {
-    const oneHour = 3600 * 1000;
-    const oneDay = 24 * oneHour;
-
-    const users = [{ uid: 'U001' }, { uid: 'U002' }, { uid: 'U003' }];
-
-    const events = [
-      { id: 'E001', start: new Date(Date.now() + oneDay + (oneHour / 2)) }, // Should create a notification
-      { id: 'E002', start: new Date(Date.now() + oneDay + oneDay) }, // Should not create a notification
-      { id: 'E003', start: new Date(Date.now() - oneDay) }, // Should not create a notification
-    ];
-
-    const invitations = [
-      {
-        id: 'I001',
-        eventId: 'E001',
-        toUser: {
-          uid: 'U001'
-        },
-        status: 'accepted'
-      },
-      {
-        id: 'I002',
-        eventId: 'E001',
-        toUser: {
-          uid: 'U002'
-        },
-        status: 'pending'
-      },
-      {
-        id: 'I003',
-        eventId: 'E002',
-        toUser: {
-          uid: 'U003'
-        },
-        status: 'accepted'
-      },
-      {
-        id: 'I004',
-        eventId: 'E003',
-        toUser: {
-          uid: 'U003'
-        },
-        status: 'accepted'
-      },
-      {
-        id: 'I005',
-        eventId: 'E001',
-        fromUser: {
-          uid: 'U003'
-        },
-        mode: 'request',
-        status: 'accepted'
-      }
-    ]
 
     // Load our test set
-    await populate('users', users);
-    await populate('events', events);
-    await populate('invitations', invitations);
+    await populate('users', data.users);
+    await populate('events', data.events);
+    await populate('invitations', data.invitations);
 
     await createNotificationsForEventsToStart();
